Add and Cancel Active Challenge working
